+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 0.3.0 / 2020-02-27
 ==================
 
+  * fix: Add local for service port
+  * tech: Add outputs on example
   * feat: Add deployment template variables, outputs
   * feat: Add more outputs